split out configuratin message
