+++ daemon/src/services/garmin/garmindeviceinformationmessage.cpp
@@ -39,6 +39,37 @@ void GarminDeviceInformationMessage::parse(const QByteArray& data)
     auto s3 = readLengthPrefixedString(data.mid(off), consumed);
     msg.deviceModel = s3.value;
 
+    // Generate Response
+    QByteArray response = generateResponse(msg);
+    response = wrapInGfdiEnvelope(5000,response);
     // Now update the device via the communicator
-    if (mCommunicator) mCommunicator->onDeviceInformationReceived(msg);
+    if (mCommunicator) {
+        mCommunicator->sendMessage("DEVICEINFORMATION RESPONSE",response);
+        mCommunicator->onDeviceInformationReceived(msg);
+    }
+}
+
+QByteArray GarminDeviceInformationMessage::generateResponse(const DeviceInformationMessage& incoming)
+{
+    QByteArray r;
+    writeU16le(r, 5024);                             // original DEVICE_INFORMATION
+    r.append(char(quint8(Status::Ack)));            // status
+
+    writeU16le(r, 150);                              // protocol version 1.50
+    writeU16le(r, 0xFFFF);                           // product number (-1 = 0xFFFF for phone)
+    writeU32le(r, 0xFFFFFFFFu);                      // our unit number
+    writeU16le(r, 7791);                             // software version  (7791 = version 77.91, matching Gadgetbridge)
+    writeU16le(r, 0xFFFF);                           // our max packet size( -1 = 0xFFFF means no limit)
+    // Bluetooth name (null-terminated)
+    r.append("Jolla-Amazfish"); r.append(char(0));
+    // Device manufacturer (null-terminated)
+    r.append("Jolla");      r.append(char(0));
+    // Device model (null-terminated)
+    r.append("SailfishOS");             r.append(char(0));
+
+     // Protocol flags (1 for v1.x, 0 for v2.x)
+    const quint8 protocolFlags = (incoming.protocolVersion / 100 == 1) ? 1 : 0;
+    r.append(char(protocolFlags));
+
+    return r;
 }
